Add error handling in the sampler.
Previously, an error in the likelihood function would abort the entire
MCMC calculation.  This can be maddening to track down because often
such errors only occur for a small set of parameters, not known to the
user, and the sampler only finds them after thousands of iterations.

Now, we catch those errors, convert the error message to an
informational message, and set the proposal log-posterior to -Inf (which
causes it to be automatically rejected).  If debugging information is
being returned, we also return a vector of flags indicating which
proposal parameters had errors, which allows a user to go back and track
down which parameters caused problems and why.